requirements.txt
Flask
Flask-SQLAlchemy
flask_marshmallow
marshmallow-sqlalchemy
mysql-connector-python
